requirements.txt
PyQt5==5.15.7
python-docx==0.8.11
pyautogui
qt-material